tutorials/requirements.txt
protobuf>=3.5.0,<3.20
grpcio==1.41.0
numpy==1.21.4
Pillow==9.1.1
torch==1.11.0
transformers==4.18.0
tritonclient[grpc,http]==2.22.3
onnxruntime==1.8.0
attrdict==2.0.1
